# apps/web/src/lib/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/tasks/{id}": {
      "get": {
        "operationId": "TasksController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A task with its dependencies, checklist, links and revisions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetail"
                }
              }
            }
          }
        },
        "tags": [
          "tasks"
        ]
      },
      "patch": {
        "operationId": "TasksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTask"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated task (manage scope, or the creator of an unassigned new request); assignee and department need assign scope, the request scope client scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetail"
                }
              }
            }
          }
        },
        "tags": [
          "tasks"
        ]
      }
    },
...
    "/api/tasks/{id}/archive": {
      "post": {
        "operationId": "TasksController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The archived task (scope all only)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetail"
                }
              }
            }
          }
        },
        "tags": [
          "tasks"
        ]
      }
    },
...
    "/api/tasks/{id}/checklist": {
      "post": {
        "operationId": "TaskPartsController_addChecklistItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskChecklistItem"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The new checklist item, last in the list (assignee, department manager or manage scope)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskChecklistItem"
                }
              }
            }
          }
        },
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/tasks/{id}/checklist/order": {
      "put": {
        "operationId": "TaskPartsController_reorderChecklist",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskChecklistOrder"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The checklist in its new order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskChecklist"
                }
              }
            }
          }
        },
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/tasks/{id}/checklist/{itemId}": {
      "patch": {
        "operationId": "TaskPartsController_updateChecklistItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskChecklistItem"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The renamed, ticked or unticked item",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskChecklistItem"
                }
              }
            }
          }
        },
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/tasks/{id}/checklist/{itemId}/archive": {
      "post": {
        "operationId": "TaskPartsController_archiveChecklistItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The item is removed from the checklist"
          }
        },
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/tasks/{id}/links": {
      "post": {
        "operationId": "TaskPartsController_addLink",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskLink"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The new link (assignee, department manager or manage scope)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskLink"
                }
              }
            }
          }
        },
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/tasks/{id}/links/{linkId}/archive": {
      "post": {
        "operationId": "TaskPartsController_archiveLink",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "linkId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The link is removed from the task"
          }
        },
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/tasks/{id}/comments": {
      "get": {
        "operationId": "TaskCommentsController_list",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 50,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Comments, oldest first; removed ones stay in place without their body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCommentPage"
                }
              }
            }
          }
        },
        "tags": [
          "tasks"
        ]
      },
      "post": {
        "operationId": "TaskCommentsController_create",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCommentInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The new comment; mentions are `@{userId}` tokens of active users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskComment"
                }
              }
            }
          }
        },
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/tasks/{id}/comments/{commentId}": {
      "patch": {
        "operationId": "TaskCommentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCommentInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The edited comment (its author only)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskComment"
                }
              }
            }
          }
        },
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/tasks/{id}/comments/{commentId}/archive": {
      "post": {
        "operationId": "TaskCommentsController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The comment is removed (its author, or `tasks.manage` with scope all)"
          }
        },
        "tags": [
          "tasks"
        ]
      }
    },
...
  },
  "info": {
    "title": "Vertex Hub API",
    "description": "",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "TaskType": {
        "type": "string",
        "enum": [
          "work",
          "client_request"
        ]
      },
...
      "TaskDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/TaskType"
          },
          "department": {
            "$ref": "#/components/schemas/DepartmentCode"
          },
          "assignee": {
            "nullable": true,
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "name": {
                "type": "string"
              },
              "archived": {
                "type": "boolean"
              },
              "inDepartment": {
                "type": "boolean"
              }
            },
            "required": [
              "id",
              "name",
              "archived",
              "inDepartment"
            ],
            "additionalProperties": false
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "priority": {
            "$ref": "#/components/schemas/TaskPriority"
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
          },
          "dueTime": {
            "nullable": true,
            "type": "string",
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$"
          },
          "overdue": {
            "type": "boolean"
          },
          "blocked": {
            "type": "boolean"
          },
          "client": {
            "nullable": true,
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "name": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "name"
            ],
            "additionalProperties": false
          },
          "project": {
            "nullable": true,
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "name": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "name"
            ],
            "additionalProperties": false
          },
          "milestone": {
            "nullable": true,
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "name": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "name"
            ],
            "additionalProperties": false
          },
          "retainer": {
            "nullable": true,
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "name": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "name"
            ],
            "additionalProperties": false
          },
          "cycle": {
            "nullable": true,
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "periodStart": {
                "type": "string",
                "format": "date",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
              },
              "periodEnd": {
                "type": "string",
                "format": "date",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
              }
            },
            "required": [
              "id",
              "periodStart",
              "periodEnd"
            ],
            "additionalProperties": false
          },
          "cycleLine": {
            "nullable": true,
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "kind": {
                "$ref": "#/components/schemas/DeliverableKind"
              },
              "label": {
                "nullable": true,
                "type": "string"
              }
            },
            "required": [
              "id",
              "kind",
              "label"
            ],
            "additionalProperties": false
          },
          "checklist": {
            "type": "object",
            "properties": {
              "done": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "total": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              }
            },
            "required": [
              "done",
              "total"
            ],
            "additionalProperties": false
          },
          "revisions": {
            "type": "object",
            "properties": {
              "clientCount": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "limit": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              }
            },
            "required": [
              "clientCount",
              "limit"
            ],
            "additionalProperties": false
          },
          "overLimitPending": {
            "type": "boolean"
          },
          "brief": {
            "nullable": true,
            "type": "string"
          },
          "needsClientApproval": {
            "type": "boolean"
          },
          "clientRequest": {
            "nullable": true,
            "type": "object",
            "properties": {
              "contact": {
                "nullable": true,
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "name": {
                    "type": "string"
                  },
                  "archived": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "id",
                  "name",
                  "archived"
                ],
                "additionalProperties": false
              },
              "requestedOn": {
                "type": "string",
                "format": "date",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
              },
              "scope": {
                "$ref": "#/components/schemas/RequestScope"
              },
              "extraWork": {
                "nullable": true,
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "title": {
                    "type": "string"
                  },
                  "billingStatus": {
                    "$ref": "#/components/schemas/ExtraWorkBilling"
                  }
                },
                "required": [
                  "id",
                  "title",
                  "billingStatus"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "contact",
              "requestedOn",
              "scope",
              "extraWork"
            ],
            "additionalProperties": false
          },
          "dependencies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskDependency"
            }
          },
          "dependents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskDependency"
            }
          },
          "checklistItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskChecklistItem"
            }
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskLink"
            }
          },
          "revisionHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskRevision"
            }
          },
          "createdBy": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "name": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "name"
            ],
            "additionalProperties": false
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "startedAt": {
            "nullable": true,
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "deliveredAt": {
            "nullable": true,
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "cancelledAt": {
            "nullable": true,
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "cancelReason": {
            "nullable": true,
            "type": "string"
          },
          "archivedAt": {
            "nullable": true,
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "readOnly": {
            "type": "boolean"
          },
          "permissions": {
            "$ref": "#/components/schemas/TaskPermissions"
          },
          "allowedTransitions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskStatus"
            }
          }
        },
        "required": [
          "id",
          "title",
          "type",
          "department",
          "assignee",
          "status",
          "priority",
          "dueDate",
          "dueTime",
          "overdue",
          "blocked",
          "client",
          "project",
          "milestone",
          "retainer",
          "cycle",
          "cycleLine",
          "checklist",
          "revisions",
          "overLimitPending",
          "brief",
          "needsClientApproval",
          "clientRequest",
          "dependencies",
          "dependents",
          "checklistItems",
          "links",
          "revisionHistory",
          "createdBy",
          "createdAt",
          "updatedAt",
          "startedAt",
          "deliveredAt",
          "cancelledAt",
          "cancelReason",
          "archivedAt",
          "readOnly",
          "permissions",
          "allowedTransitions"
        ],
        "additionalProperties": false
      },
      "RequestScope": {
        "type": "string",
        "enum": [
          "in_scope",
          "out_of_scope"
        ]
      },
...
      "CreateTaskChecklistItem": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        },
        "required": [
          "text"
        ]
      },
...
    }
  }
}